functions/requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
google-cloud-vision
google-cloud-storage
Pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
PyPDF2>=3.0.0
requests>=2.31.0
